Library

Dubins.jl Library

Dubins.DubinsPath — Type.

The data structure that holds the full dubins path.

Its data fields are as follows:

the initial configuration, qi,

the params vector that contains the length of each segment, params,

the turn-radius, ρ, and,

the Dubins path type given by the @enum DubinsPathType

source

Dubins.DubinsPath — Method.

Empty constructor for the DubinsPath type

source

Dubins.dubins_extract_subpath — Method.

Convenience function to extract a sub-path

path - an initialized path

t - a length measure, where 0 < t < dubins_path_length(path)

newpath - the resultant path

return - zero on successful completion

source

Dubins.dubins_path — Method.

Generate a path with a specified word from an initial configuratioon to a target configuration, with a specified turning radius

path - resultant path

q0 - a configuration specified by a 3-element vector x, y, theta

q1 - a configuration specified by a 3-element vector x, y, theta

rho - turning radius of the vehicle

path_type - the specified path type to use

return - error code

source

Dubins.dubins_path_endpoint — Method.

Convenience function to identify the endpoint of a path

path - an initialized path

q - the end point

return - zero on successful completion

source

Dubins.dubins_path_length — Method.

Calculate the length of an initialized path

path - path to find the length of

return - path length

source

Dubins.dubins_path_sample — Method.

Calculate the configuration along the path, using the parameter t

path - an initialized path

t - length measure where 0 <= t < dubins_path_length(path)

q - the configuration result [x, y, θ]

return - non-zero error code if 't' is not in the correct range

source

Dubins.dubins_path_sample_many — Method.

Walk along the path at a fixed sampling interval, calling the callback function at each interval

The sampling process continues until the whole path is sampled, or the callback returns a non-zero value

path - the path to sample

step_size - the distance along the path for subsequent samples

cb - the callback function to call for each sample

return - error code

source

Dubins.dubins_path_type — Method.

Extract the integer that represents which path type was used

path - an initialized path

return - one of LSL-0, LSR-1, RSL-2, RSR-3, RLR-4, LRL-5

source

Dubins.dubins_segment_length — Method.

Calculate the length of a specific segment of an initialized path

path - path to find the length of

i - the segment for which the length is required (1-3)

return - segment length

source

Dubins.dubins_segment_length_normalized — Method.

Calculate the normalized length of a specific segment of an initialized path

path - path to find the length of

i - the segment for which the length is required (1-3)

return - normalized segment length

source

Dubins.dubins_shortest_path — Method.

Generate a path from an initial configuration to a target configuration with a specified maximum turning radius

A configuration is given by [x, y, θ], where θ is in radians,

path - resultant path

q0 - a configuration specified by a 3-element vector [x, y, θ]

q1 - a configuration specified by a 3-element vector [x, y, θ]

rho - turning radius of the vehicle

return - error code

source

Dubins.DubinsIntermediateResults — Type.

This data structure holds the information to compute the Dubins path in the transformed coordinates where the initial (x,y) is translated to the origin, the final the coordinate axis is rotated to make the x-axis aligned with the line joining the two points. The variable names follow the convention used in the paper "Classification of the Dubins set" by Andrei M. Shkel and Vladimir Lumelsky

source

Dubins.DubinsIntermediateResults — Method.

Empty constructor for the DubinsIntermediateResults data type

source

Dubins.dubins_segment — Method.

Operators that transform an arbitrary point qi, [x, y, θ], into an image point.

The three operators correspond to L_SEG, R_SEG, and S_SEG

L_SEG(x, y, θ, t) = [x, y, θ] + [ sin(θ + t) - sin(θ), -cos(θ + t) + cos(θ), t]

R_SEG(x, y, θ, t) = [x, y, θ] + [-sin(θ - t) + sin(θ), cos(θ - t) - cos(θ), -t]

S_SEG(x, y, θ, t) = [x, y, θ] + [ cos(θ) * t, sin(θ) * t, 0]

source

Dubins.dubins_word — Method.

The function to call the corresponding Dubins path based on the path_type

return - error code; zero on successful completion

source

Dubins.set_tolerance — Method.

Reset tolerance value

source